增加删除条件

+++ pytest_project/clear_download_dir.py
@@ -5,7 +5,10 @@ download_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'downloa
 
 def clear_download_files():
     for i in os.listdir(download_dir):
-        os.remove(download_dir + os.sep + i)
+        if i == '__init__.py':
+            break
+        else:
+            os.remove(download_dir + os.sep + i)
 
 
 clear_download_files()
